server/pyproject.toml
[tool.poetry]
name = "zipzoong"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
isort = "^5.10.1"
flake8 = "^4.0.1"
black = "^22.3.0"
drf-yasg = "^1.20.0"
django-dotenv = "^1.4.2"
mysqlclient = "^2.1.1"
PyMySQL = "^1.0.2"
django-cors-headers = "^3.13.0"
django-ninja = "^0.19.1"
django-ninja-extra = "^0.15.8"
ninja-schema = "^0.13.0"
PyJWT = "^2.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
